requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
statsmodels>=0.14
dowhy>=0.11
econml>=0.15
PyYAML>=6.0
fastapi>=0.110
uvicorn>=0.23
prometheus-client>=0.16
pydantic>=2.0
neo4j>=5.14
kafka-python>=2.0
